General refactoring. You can now fetch info for a single package.
